rename dynamic_swin_transformer
