+++ configs/detection/dino/dino_4scale_swinl.py
@@ -3,7 +3,7 @@ model = dict(
     type='Detection',
     pretrained='/root/pretrain/warpper_swin_large_patch4_window12_384_22k.pth',
     backbone=dict(
-        type='DINOSwinTransformer',
+        type='SwinTransformer',
         pretrain_img_size=384,
         embed_dim=192,
         depths=[2, 2, 18, 2],
