+++ easycv/models/backbones/__init__.py
@@ -17,6 +17,5 @@ from .resnet import ResNet
 from .resnet_jit import ResNetJIT
 from .resnext import ResNeXt
 from .shuffle_transformer import ShuffleTransformer
-from .swin_transformer import DINOSwinTransformer
-from .swin_transformer_dynamic import SwinTransformer
+from .swin_transformer import SwinTransformer
 from .vitdet import ViTDet
